\name{get_nominal_split}
\alias{get_nominal_split}
\title{Binary partitions of a nominal variable}
\usage{
  get_nominal_split(v)
}
\arguments{
  \item{v}{vector containing the nominal categories}
}
\description{
  Internal function. \code{get_nominal_split} is called by
  \code{pathmox}
}
\keyword{internal}

