\name{get_ordinal_split}
\alias{get_ordinal_split}
\title{Binary partitions of an ordinal variable}
\usage{
  get_ordinal_split(cat.exe)
}
\arguments{
  \item{cat.exe}{vector of categories}
}
\description{
  Internal function. \code{get_ordinal_split} is called by
  \code{pathmox}
}
\keyword{internal}

