\name{pathmox-package}
\alias{pathmox-package}
\docType{package}
\title{
  Package of Segmentation Trees in Partial Least Squares Path Modeling}
\description{
  Contains functions of PATHMOX and TECHMOX algorithms that perform segmentation trees in PLS Path Modeling}
\details{
\tabular{ll}{
  Package: \tab pathmox\cr
  Type: \tab Package\cr
  Version: \tab 0.1-1\cr
  Date: \tab 2012-03-14\cr
  Depends: \tab plspm\cr
  License: \tab GPL (>= 2)\cr
  LazyLoad: \tab yes\cr
  }
  The main functions are \code{\link{pathmox}}, and \code{\link{techmox}}. 
}
\author{
  Authors: Gaston Sanchez, Tomas Aluja

  Maintainer: Gaston Sanchez \email{plspmr@gmail.com}
}
\references{
  Sanchez, G. (2009) \emph{PATHMOX Approach: Segmentation Trees in Partial Least Squares Path Modeling.} Doctoral Dissertation. Departament Estadistica i Investigacio Operativa, Universitat Politecnica de Catalunya.

  \url{http://www.plsmodeling.com}
}
\keyword{ package }
