\name{plot.bootnodes}
\alias{plot.bootnodes}
\title{Plot bootstrap results of terminal nodes}
\usage{
  plot.bootnodes(x, pc = 1, ...)
}
\arguments{
  \item{x}{object of class \code{"bootnodes"}}

  \item{pc}{integer indicating the number of path
  coefficient to be plotted}

  \item{\dots}{Further arguments are ignored}
}
\description{
  This function is deprecated Plot method for objects of
  class \code{"bootnodes"}

  The function \code{plot.bootnodes} displays the value of
  the selected path coefficient for the root node and the
  terminal nodes of PATHMOX and TECHMOX trees. \cr In
  addition, the value of the mean bootstrap as well as the
  values of the confidence interval are also shown.
}

