\name{print.treemox.pls}
\alias{print.treemox.pls}
\title{Printing treemox.pls objects}
\description{
  Printing method for objects of class \code{"treemox.pls"}
}
\usage{
  \method{print}{treemox.pls}(x, \dots)
}
\arguments{
  \item{x}{Object of class \code{"treemox.pls"} returned by \code{\link{treemox.pls}}.}
  \item{\dots}{Further arguments are ignored.}
}
\details{
  The function \code{print.treemox.pls} displays basic PLS-PM results for the nodes of pathmox and techmos trees. 
}
\seealso{\code{\link{pathmox}}, \code{\link{techmox}}, \code{\link{treemox.pls}}}
