\name{test_sanchez_aluja}
\alias{test_sanchez_aluja}
\title{Adaptation by Sanchez-Aluja of Lebart test}
\usage{
  test_sanchez_aluja(Y.lvs, IDM, split.a, split.b)
}
\arguments{
  \item{Y.lvs}{matrix with scores of latent variables}

  \item{IDM}{inner design matrix}

  \item{split.a}{split a (left side)}

  \item{split.b}{split b (right side)}
}
\description{
  Internal function. \code{test_sanchez_aluja} is called by
  \code{pathmox}
}
\keyword{internal}

