\name{pathmox-package}
\alias{pathmox-package}
\docType{package}
\title{
  Pathmox Approach: Segmentation Trees in Partial Least Squares Path Modeling}
\description{
  pathmox is the cousin package of the awesome package plspm;
  pathmox provides a very interesting solution for handling segmentation
  variables in PLS Path Modeling: segmentation trees in PLS Path Modeling}
\details{
\tabular{ll}{
  Package: \tab pathmox\cr
  Type: \tab Package\cr
  Version: \tab 0.1.5\cr
  Date: \tab 2012-11-17\cr
  Depends: \tab R (>=2.15.1), plspm\cr
  Suggests: \tab plsdepot\cr
  License: \tab GPL-3\cr
  LazyLoad: \tab yes\cr
  }
}
\author{
  Authors: Gaston Sanchez, Tomas Aluja

  Maintainer: Gaston Sanchez \email{gaston.stat@gmail.com}
}
\references{
  \url{http://www.gastonsanchez.com/pathmox}
  
  Sanchez, G. (2009) \emph{PATHMOX Approach: Segmentation Trees in Partial Least Squares Path Modeling.} Doctoral Dissertation.
  \url{http://www.gastonsanchez.com/thesis-pathmox}
}
\keyword{ package }
